# api/openapi.yaml
openapi: 3.0.3
info:
  title: YKAY Virtual School API
  version: 1.0.0
  description: |
    Full commercial, SEO-first, curriculum-governed tutoring marketplace and virtual school.
    Built to Tuteria-level parity + beyond (blog, B2B, referrals, escrow, vetting).
servers:
  - url: http://localhost:8080/api/v1
    description: Local
  - url: https://api.ykay.com/api/v1
    description: Production
...
paths:
  /auth/register:
    post:
      tags: [identity]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, roles]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                roles: { type: array, items: { type: string, enum: [STUDENT, PARENT, TUTOR] } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/login:
    post:
      tags: [identity]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
...
  /programmes:
    get:
      tags: [programmes]
      summary: List programmes (filter by curriculum, level, exam, format)
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: curriculum, in: query, schema: { type: string } }
        - { name: level, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { type: string, enum: [COHORT, PRIVATE, BOOTCAMP] } }
      responses:
        '200':
          description: Programmes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProgrammes'
...
components:
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        status: { type: string }
        timezone: { type: string }
...
    PaginatedProgrammes:
      type: object
      properties:
        data: { type: array, items: { type: object } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total_items: { type: integer }
        total_pages: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
...